Added statusline settings. Fixed indentation
I want to use a custom statusline for both Gundo splits so i added two
variables for it setting it. If the variables are empty Gundo works like
normal but if there is something in them the text get used. Example:
  let g:gundo_preview_statusline = "Gundo Preview"
  let g:gundo_tree_statusline = "Gundo Tree"

The indentation was off in a few places so i fixed it. The real tabs was
removed and replaced with spaces like the rest of the source file.